+++ plugins/Moodle/Custom/CreateParents.php
@@ -85,6 +85,12 @@ function core_user_create_users_response( $response )
 	)
 	 */
 
+	if ( empty( $response[0]['id'] ) )
+	{
+		// Fix SQL error when no ID returned.
+		return null;
+	}
+
 	DBQuery( "INSERT INTO MOODLEXROSARIO (\"column\", rosario_id, moodle_id)
 		VALUES('staff_id', '" . $id . "', " . $response[0]['id'] . ")" );